+++ src/logger/console.logger.ts
@@ -1,7 +1,6 @@
 import type { Logger } from './Logger.types.js';
 
 export const consoleLogger: Logger = {
-	trace: console.debug,
 	debug: console.debug,
 	info: console.info,
 	warn: console.warn,